AkarinForge 20180702-nightly

+++ projects/Forge/src/main/java/net/minecraft/entity/ai/EntityAIEatGrass.java
@@ -78,7 +78,7 @@ public class EntityAIEatGrass extends EntityAIBase
 
             if (IS_TALL_GRASS.apply(this.entityWorld.getBlockState(blockpos)))
             {
-                if (net.minecraftforge.event.ForgeEventFactory.getMobGriefingEvent(this.entityWorld, this.grassEaterEntity))
+                if (!org.bukkit.craftbukkit.event.CraftEventFactory.callEntityChangeBlockEvent(this.grassEaterEntity, this.grassEaterEntity.world.getWorld().getBlockAt(blockpos.getX(), blockpos.getY(), blockpos.getZ()), org.bukkit.Material.AIR, !this.entityWorld.getGameRules().getBoolean("mobGriefing")).isCancelled() && net.minecraftforge.event.ForgeEventFactory.getMobGriefingEvent(this.entityWorld, this.grassEaterEntity)) // CraftBukkit
                 {
                     this.entityWorld.destroyBlock(blockpos, false);
                 }
@@ -91,7 +91,7 @@ public class EntityAIEatGrass extends EntityAIBase
 
                 if (this.entityWorld.getBlockState(blockpos1).getBlock() == Blocks.GRASS)
                 {
-                    if (net.minecraftforge.event.ForgeEventFactory.getMobGriefingEvent(this.entityWorld, this.grassEaterEntity))
+                    if (!org.bukkit.craftbukkit.event.CraftEventFactory.callEntityChangeBlockEvent(this.grassEaterEntity, this.grassEaterEntity.world.getWorld().getBlockAt(blockpos1.getX(), blockpos1.getY(), blockpos1.getZ()), org.bukkit.Material.AIR, !this.entityWorld.getGameRules().getBoolean("mobGriefing")).isCancelled() && net.minecraftforge.event.ForgeEventFactory.getMobGriefingEvent(this.entityWorld, this.grassEaterEntity)) // CraftBukkit // Akarin Forge - corrects position
                     {
                         this.entityWorld.playEvent(2001, blockpos1, Block.getIdFromBlock(Blocks.GRASS));
                         this.entityWorld.setBlockState(blockpos1, Blocks.DIRT.getDefaultState(), 2);
