+++ projects/Forge/src/main/java/net/minecraft/entity/ai/EntityAIHarvestFarmland.java
@@ -61,7 +61,7 @@ public class EntityAIHarvestFarmland extends EntityAIMoveToBlock
 
             if (this.currentTask == 0 && block instanceof BlockCrops && ((BlockCrops)block).isMaxAge(iblockstate))
             {
-                world.destroyBlock(blockpos, true);
+                if (!org.bukkit.craftbukkit.event.CraftEventFactory.callEntityChangeBlockEvent(this.villager, blockpos, Blocks.AIR, 0).isCancelled()) world.destroyBlock(blockpos, true); // CraftBukkit
             }
             else if (this.currentTask == 1 && iblockstate.getMaterial() == Material.AIR)
             {
@@ -74,32 +74,39 @@ public class EntityAIHarvestFarmland extends EntityAIMoveToBlock
 
                     if (!itemstack.isEmpty())
                     {
+                        Block planted = null; // CraftBukkit
+                        IBlockState state = null; // Akarin Forge - respect Forge
                         if (itemstack.getItem() == Items.WHEAT_SEEDS)
                         {
-                            world.setBlockState(blockpos, Blocks.WHEAT.getDefaultState(), 3);
+                            planted = Blocks.WHEAT; // CraftBukkit
                             flag = true;
                         }
                         else if (itemstack.getItem() == Items.POTATO)
                         {
-                            world.setBlockState(blockpos, Blocks.POTATOES.getDefaultState(), 3);
+                            planted = Blocks.POTATOES; // CraftBukkit
                             flag = true;
                         }
                         else if (itemstack.getItem() == Items.CARROT)
                         {
-                            world.setBlockState(blockpos, Blocks.CARROTS.getDefaultState(), 3);
+                            planted = Blocks.CARROTS; // CraftBukkit
                             flag = true;
                         }
                         else if (itemstack.getItem() == Items.BEETROOT_SEEDS)
                         {
-                            world.setBlockState(blockpos, Blocks.BEETROOTS.getDefaultState(), 3);
+                            planted = Blocks.BEETROOTS; // CraftBukkit
                             flag = true;
                         }
                         else if (itemstack.getItem() instanceof net.minecraftforge.common.IPlantable) {
                             if(((net.minecraftforge.common.IPlantable)itemstack.getItem()).getPlantType(world,blockpos) == net.minecraftforge.common.EnumPlantType.Crop) {
-                                world.setBlockState(blockpos, ((net.minecraftforge.common.IPlantable)itemstack.getItem()).getPlant(world,blockpos),3);
+                                state = ((net.minecraftforge.common.IPlantable) itemstack.getItem()).getPlant(world,blockpos); // Akarin Forge - respect Forge
                                 flag = true;
                             }
                         }
+                        // CraftBukkit start
+                        if (planted != null && !org.bukkit.craftbukkit.event.CraftEventFactory.callEntityChangeBlockEvent(this.villager, blockpos, planted, 0).isCancelled()) {
+                            world.setBlockState(blockpos, planted == null ? state : planted.getDefaultState(), 3); // Akarin Forge - respect Forge
+                        } else flag = false;
+                        // CraftBukkit end
                     }
 
                     if (flag)
